feat: clean up code, update types

+++ features/create-vault/utils/format-data.ts
@@ -1,5 +1,5 @@
-import { VaultFactoryArgs } from '../../../types';
-import { VAULT_TOTAL_BASIS_POINTS } from 'consts/vault-hub';
+import { VaultFactoryArgs } from 'types';
+import { VAULT_TOTAL_BASIS_POINTS } from 'modules/vaults/consts';
 import { CreateVaultSchema } from '../create-vault-form/create-vault-form-context/validation';
 
 export const formatCreateVaultData = (
@@ -8,8 +8,9 @@ export const formatCreateVaultData = (
   const { nodeOperator, ...payload } = values;
   const { confirmExpiry, nodeOperatorFeeBP } = payload;
   const confirmExpiryFormatted = BigInt(confirmExpiry * 60 * 60);
-  const nodeOperatorFeeBPFormatted =
-    (nodeOperatorFeeBP * VAULT_TOTAL_BASIS_POINTS) / 100;
+  const nodeOperatorFeeBPFormatted = BigInt(
+    (nodeOperatorFeeBP * VAULT_TOTAL_BASIS_POINTS) / 100,
+  );
 
   return {
     ...payload,
